Fix: Renamed module structure
