+++ src/yoru/extensions.js
@@ -1,0 +1,4 @@
+// Import helper for module Extensions
+
+import { YoruArray } from 'yoru/extensions/_all';
+export { YoruArray };
