+++ src/yoru/extensions/_all.js
@@ -1,0 +1,6 @@
+//
+// 夜/Extensions
+//
+
+import YoruArray from 'yoru/extensions/array';
+export { YoruArray };
