+++ src/yoru/extensions/array.js
@@ -1,5 +1,5 @@
 //
-// 夜/追加/array.js
+// 夜/Extensions/Array
 //
 
 /**
