+++ src/yoru/internals/listener.js
@@ -1,5 +1,5 @@
 //
-// 夜/心/listener.js
+// 夜/Internals/Listener
 //
 
 import YoruObject from 'yoru/object';
