--- src/yoru/kage.js
@@ -1,4 +1,0 @@
-// Import helper for module Kage
-
-import { ShadowMaker, TemplateConsumer } from 'yoru/kage/_all';
-export { ShadowMaker, TemplateConsumer };
