--- src/yoru/kage/_all.js
@@ -1,8 +1,0 @@
-//
-// 夜/影
-//
-
-import ShadowMaker from 'yoru/kage/shadow-maker';
-import TemplateConsumer from 'yoru/kage/template-consumer';
-
-export { ShadowMaker, TemplateConsumer };
