--- src/yoru/kokoro.js
@@ -1,4 +1,0 @@
-// Import helper for module Kokoro
-
-import { Listener, Preloader } from 'yoru/kokoro/_all';
-export { Listener, Preloader };
