--- src/yoru/kokoro/_all.js
@@ -1,8 +1,0 @@
-//
-// 夜/心
-//
-
-import Listener from 'yoru/kokoro/listener';
-import Preloader from 'yoru/kokoro/preloader';
-
-export { Listener, Preloader };
