--- src/yoru/komono.js
@@ -1,4 +1,0 @@
-// Import helper for module Komono
-
-import { Logger, Scribe, Run } from 'yoru/komono/_all';
-export { Logger, Scribe, Run };
