--- src/yoru/komono/_all.js
@@ -1,9 +1,0 @@
-//
-// 夜/小物
-//
-
-import Logger from 'yoru/komono/logger';
-import Scribe from 'yoru/komono/scribe';
-import Run from 'yoru/komono/run';
-
-export { Logger, Scribe, Run };
