+++ src/yoru/shadow.js
@@ -1,0 +1,4 @@
+// Import helper for module Shadow
+
+import { ShadowMaker, TemplateConsumer } from 'yoru/shadow/_all';
+export { ShadowMaker, TemplateConsumer };
