+++ src/yoru/shadow/_all.js
@@ -1,0 +1,8 @@
+//
+// 夜/Shadow
+//
+
+import ShadowMaker from 'yoru/shadow/shadow-maker';
+import TemplateConsumer from 'yoru/shadow/template-consumer';
+
+export { ShadowMaker, TemplateConsumer };
