--- src/yoru/tsuika.js
@@ -1,4 +1,0 @@
-// Import helper for module Tsuika
-
-import { YoruArray } from 'yoru/tsuika/_all';
-export { YoruArray };
