--- src/yoru/tsuika/_all.js
@@ -1,6 +1,0 @@
-//
-// 夜/追加
-//
-
-import YoruArray from 'yoru/tsuika/array';
-export { YoruArray };
